- Stacked cards and cards in discard affect game ONLY when their side (Shadow/FP) matches their owner's current role.

+++ trunk/gemp-lotr/gemp-lotr-cards/src/main/java/com/gempukku/lotro/cards/set15/orc/Card15_117.java
@@ -41,7 +41,7 @@ public class Card15_117 extends AbstractMinion {
 
     @Override
     public Modifier getAlwaysOnModifier(PhysicalCard self) {
-        return new KeywordModifier(self, self, null, Keyword.DAMAGE, new CountActiveEvaluator(5, null, CardType.COMPANION));
+        return new KeywordModifier(self, self, null, Keyword.DAMAGE, new CountActiveEvaluator(5, (Integer) null, CardType.COMPANION));
     }
 
     @Override
